requirements.txt
Flask==3.0.0
pytubefix
openai
python-dotenv